+++ common/build.gradle.kts
@@ -1,22 +1,17 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
-    alias(libs.plugins.hilt)
-    kotlin("kapt")
 }
 
 android {
-    namespace = "com.test.base"
+    namespace = "com.example.common"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.test.base"
         minSdk = 24
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -35,9 +30,6 @@ android {
     kotlinOptions {
         jvmTarget = "11"
     }
-    viewBinding {
-        enable = true
-    }
 }
 
 dependencies {
@@ -45,20 +37,9 @@ dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
-    implementation(libs.androidx.activity)
-    implementation(libs.androidx.constraintlayout)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
 
-    kapt(libs.hilt.compiler)
-    implementation(libs.hilt)
-    implementation(libs.bundles.networking)
-
-    implementation(libs.bundles.navigator)
-
-    implementation(project(":data"))
-    implementation(project(":presentation"))
-    implementation(project(":domain"))
-    implementation(project(":common"))
+    implementation(libs.logger)
 }
