+++ data/build.gradle.kts
@@ -44,6 +44,7 @@ dependencies {
     androidTestImplementation(libs.androidx.espresso.core)
 
     implementation(project(":domain"))
+    implementation(project(":common"))
 
     implementation(libs.hilt)
     kapt(libs.hilt.compiler)
