+++ domain/build.gradle.kts
@@ -13,3 +13,7 @@ kotlin {
         jvmTarget = org.jetbrains.kotlin.gradle.dsl.JvmTarget.JVM_11
     }
 }
+
+dependencies {
+    implementation(project(":common"))
+}
